Added boolean arguments to the default bus name request functions
The undocumented argument `flags` was removed. Since it was not
documented it is removed without going through deprecation.

The boolean flags provide the control over the name acquisition
behavior.

Looks like the `request_default_bus_name` was async on accident.
Add a special deprecation warning on awaiting the return object.