Add SdBusMessage.parse_to_tuple() method
Parses the message data to a tuple. The main difference is handling
of no data and a single complete type messages. When message has
no data returns a zero size tuple. When message has a single complete
type return a tuple of one element.

This makes it simpler to implement the D-Bus -> Python calls
because now all it takes is calling Python function with unpacked
tuple. Unpacking zero size tuple is equivalent to calling function
with no arguments.

This also fixes methods that take a single struct. Before there was
ambiguity if a method was called with a struct or multiple complete
types. Now a single struct would be a part of one element tuple.