Use ParamSpec for DbusMethodAsync instead of masking it
Originally the DbusMethodAsync was masked under the original
function meaning type checker treated it as it was the original
function and not a DbusMethodAsync.

However, it is planned to add new methods to the DbusMethodAsync
so masking it is no longer an option.

ParamSpec is only available since Python 3.10 so use the
`typing_extensions` import hidden under TYPE_CHECKING if statement.