DRAFT: Add sdbus.utils.inspect.inspect_dbus_path
Returns the D-Bus path of an object.

If called on a D-Bus proxy returns path of the proxied object. If
a bus is passed as a second argument checks that the proxy is
attached to it or raises LookupError if it is attacked to a different
bus.

IF called on a local D-Bus object returns an exported path. If an
object is not exported or exported to a different bus raises LookupError.