Add sdbus.utils.inspect.inspect_dbus_path
If called on a D-Bus proxy returns path of the proxied object.

If called on a local D-Bus object returns the exported D-Bus path.
If object is not exported raises ``LookupError``.

If called on an object that is unrelated to D-Bus raises ``TypeError``.

The object's path is inspected in the context of the given bus and if the
object is attached to a different bus the ``LookupError`` will be raised.
If the bus argument is not given or is ``None`` the default bus will be
checked against.